+++ src/main/java/de/samply/frontend/Action.java
@@ -4,6 +4,7 @@ public record Action(
         String path,
         String method,
         String[] params,
-        String explanation
+        String explanation,
+        Integer priority
 ) {
 }