+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-12-06]
+## [0.0.1 - 2024-12-09]
 ### Added
 - First version of the project
 - Spring Application
@@ -172,3 +172,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Support for Thymeleaf standard processor by converting hyphenated variables to camel case
 - Process variables recursively in Thymeleaf
 - Creator name in Frontend Project
+- Action explanation priority